Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
weasyprint==60.2
Pillow==10.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-storages==1.14.2
boto3==1.34.0
sentry-sdk==1.38.0
django-ratelimit==4.1.0
django-ipware==5.0.0

